+++ app/controllers/replies_controller.rb
@@ -10,12 +10,16 @@ class RepliesController < ApplicationController
     end
 
     post '/posts/:post_id/replies' do 
-        # binding.pry
         post = Post.find_by_id(params[:post_id])
-        reply = current_user.replies.create(params[:reply])
-        reply.post = post
-        reply.save
-        redirect "/posts/#{post.id}"
+        if !params[:reply][:content].empty?
+            reply = current_user.replies.create(params[:reply])
+            reply.post = post
+            reply.save
+            redirect "/posts/#{post.id}"
+        else
+            @error = "Your reply can't be empty"
+            erb :'/replies/new'
+        end
     end
 
 
@@ -32,14 +36,22 @@ class RepliesController < ApplicationController
     delete '/replies/:id' do 
         reply = Reply.find_by_id(params[:id])
         post_id = reply.post.id
-        reply.destroy
-        redirect "/posts/#{post_id}"
+        if reply.user == current_user
+            reply.destroy
+            redirect "/posts/#{post_id}"
+        else
+            redirect "/posts/#{post_id}"
+        end
     end 
 
-    patch '/replies/:id' do 
+    patch '/replies/:id' do
         reply = Reply.find_by_id(params[:id])
         post_id = reply.post.id
-        reply.update(params[:reply])
-        redirect "/posts/#{post_id}"
+        if reply.user == current_user
+            reply.update(params[:reply])
+            redirect "/posts/#{post_id}"
+        else
+            redirect "/posts/#{post_id}" 
+        end
     end
 end
